+++ param.go
@@ -1,0 +1,32 @@
+package httprouter
+
+// Param is a single URL parameter, consisting of a key and a value.
+type Param struct {
+	Key   string
+	Value string
+}
+
+// Params is a Param-slice, as returned by the router.
+// The slice is ordered, the first URL parameter is also the first slice value.
+// It is therefore safe to read values by the index.
+type Params []Param
+
+// DefName returns the value and true of the first Param which key matches the given name.
+// If no matching Param is found, an empty string and false is returned.
+func (ps Params) DefName(name string) (string, bool) {
+	for _, p := range ps {
+		if p.Key == name {
+			return p.Value, true
+		}
+	}
+
+	return "", false
+}
+
+// ByName returns the value of the first Param which key matches the given name.
+// If no matching Param is found, an empty string is returned.
+func (ps Params) ByName(name string) string {
+	val, _ := ps.DefName(name)
+
+	return val
+}
